Katana initial working prototype
Gradle Plugin is required for Kotlin compiler plugins, its main purpose is to allow users to pass configuration variables and then point to the correct version(s) of the compiler plugins.

The native plugin is where the majority of logic is, we are just using the compiler IR types to pull out information about the functions, types, and etc and write that out to a (json) file.

The JVM plugin doesn't actually do anything, but it exists so that if the plugin is loaded in a JVM context it will not crash the build.